stop adding the show name to an api search using a tv id
NZB Hydra would add the show name to a api search when using a TV ID.  Such as:
/api?t=tvsearch&extended=1&offset=0&limit=100&cat=5000&tvdbid=290853&ep=2&season=2&q=Fear+the+Walking+Dead

A successful search is returned on some indexers if the title is included.  However, some indexers return 0 results.  Removing the title works successfully on all indexers that support that particular ID.

Now the searches look like this:
/api?t=tvsearch&extended=1&offset=0&limit=100&cat=5000&tvdbid=290853&ep=2&season=2

This is also mirrors the format that is used when NZB Hydra tries to find out what search types an indexer supports.