Update main-stats.html
Changed the order of the last four charts (day/hour) to accommodate the tables.  If a user prefers, or would like to see, the tables, the rows for each table will correspond across the div row and each table will have the same height.

Also made the tables sortable (by clicking on the header) using some standard angular syntax.